+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Add open-telemetry trace layers to `miden-remote-prover` and `miden-remote-prover-proxy` (#1061).
 - Add open-telemetry stats for the mempool (#1073).
 - Add open-telemetry stats for the network transaction builder state (#1073).
+- RPC requests can now verify their network by setting the `miden-network` header to the genesis commitment (#1084).
 
 ### Changes